BDA Course Assignment-1/requirements.txt
pyspark==3.5.0
python-dotenv==1.0.0

